+++ test/e2e/configurer/chain/queries.go
@@ -23,6 +23,7 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
+	icacontrollertypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/controller/types"
 	"github.com/stretchr/testify/require"
 
 	"github.com/babylonlabs-io/babylon/v2/test/e2e/util"
@@ -445,6 +446,18 @@ func (n *NodeConfig) QueryTx(txHash string, overallFlags ...string) (sdk.TxRespo
 
 	txAuth := txResp.Tx.GetCachedValue().(*sdktx.Tx)
 	return txResp, txAuth
+}
+
+func (n *NodeConfig) QueryICAAccountAddress(owner, connectionID string) string {
+	path := fmt.Sprintf("ibc/apps/interchain_accounts/controller/v1/owners/%s/connections/%s", owner, connectionID)
+	bz, err := n.QueryGRPCGateway(path, url.Values{})
+	require.NoError(n.t, err)
+
+	var resp icacontrollertypes.QueryInterchainAccountResponse
+	err = util.Cdc.UnmarshalJSON(bz, &resp)
+	require.NoError(n.t, err)
+
+	return resp.Address
 }
 
 func (n *NodeConfig) WaitUntilCurrentEpochIsSealedAndFinalized(startEpoch uint64) (lastFinalizedEpoch uint64) {